+++ src/core/attack/mod.rs
@@ -138,6 +138,10 @@ pub const RECONNAISSANCE: &[Technique] = &[
         name: "Vulnerability Scanning",
     },
     Technique {
+        id: "T1595.003",
+        name: "Wordlist Scanning",
+    },
+    Technique {
         id: "T1596",
         name: "Search Open Technical Databases",
     },